Major reworking and tweaks for search! - The schema for indexing has been updated to attempt to do proper partial word indexing. Note: at the moment fragments between 3 and 100 letters are tokenised. - The way searching works has been reworked so that searches across all models are first carried out and then the required top level search result models are selected out of this. This will hopefully produce more accurate results and also now take into account annotations.
Note: there is still a lot more work to do on this in the future, and some of the changes here are proverbial hacks that won't stand the test of time! Maybe in the future we should look to using a more robust plugin instead of acts_as_solr. Such as (potentially): http://github.com/outoftime/sunspot/tree/master

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@352 1acaf488-1b9d-4964-8077-0a2fdb83ef00