Annotations plugin: implemented versioning of annotation objects!
This uses a library from the version_fu plugin - a slightly customised version of this is embedded within the annotations plugin (which won't conflict if the version_fu is installed in the main codebase).

Migrations have changed, which will require an update to the main codebase.

A script will also need to be developed, to ensure initial annotation_version records are created for existing annotation objects.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@387 1acaf488-1b9d-4964-8077-0a2fdb83ef00