Initial implementation of the tags index (ie: main tag cloud) and tag show page. So now /tags will take you to the tags index page and /tags/{tag_name} will take you to the results of items found with that tag.
The tag index page needs to be updated to show a proper tag cloud (and individual tags need to be clickable).

The tag show page shows results items found in tabs (exactly the same as for search results) - refactored the tabbed results UI into a partial that is being used by the search show page as well as the tag show page.

Now all tags in the site need to be made clickable, and links to the main tags index page need to be put in.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@421 1acaf488-1b9d-4964-8077-0a2fdb83ef00