IMPORTANT: changed development mode to not reload classes each time, as this was causing huge issues (ref: https://rails.lighthouseapp.com/projects/8994/tickets/1290 and https://rails.lighthouseapp.com/projects/8994/tickets/1339).
Unfortunately this does mean that any changes to lib files, controllers, helpers and model will require a restart of your server to see the changes take effect. I think changes in views don't require a restart. I'm afraid this is a necessary change. Apparently Rails 2.3.2 fixes the above mentioned issue, so hopefully we won't have to deal with this for too long.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@745 1acaf488-1b9d-4964-8077-0a2fdb83ef00