Implemented auto suggestions of search terms when users start typing a search term.
The suggested queries come from a flat file - data/search_query_suggestions.txt, which can be updated via a background job. The background job updates the file by taking all the categories, and all the existing queries people have searched for, and outputting it to this file, to be used as suggestions.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@879 1acaf488-1b9d-4964-8077-0a2fdb83ef00