Added caching to the search process. In particular, to the part that fetches the search docs/items from Solr. Since this part is usually the same for all searches of the same query regardless of search scope, we can optimise it so that Solr doesn't have to keep giving us back the same collection of items found (that we then process).
The cache time is based on the SEARCH_ITEMS_FROM_SOLR_CACHE_TIME setting.

This reduces the load coming in from many simultaneous searches of the same query but different scope. For example, when the BioCatalogue plugin for Taverna searches for Services, SOAP operations, REST endpoints, etc all at the same time but as separate scoped requests.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@2187 1acaf488-1b9d-4964-8077-0a2fdb83ef00