Service monitoring: test with HEAD and retry services.
Switch the default method used for testing to HEAD so that services do not have
to produce a document when pinged.

If a 405 is returned then read the Allow header and try again with one of the
methods it allows (restricted to OPTIONS and GET).

If a 500 is returned, which is quite common for some WSDL endpoints, retry with
OPTIONS and then GET.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@2564 1acaf488-1b9d-4964-8077-0a2fdb83ef00