Add more status codes to availability checking code.
Now treats 401 as success. This is a little contentious but it is hard to argue
that a 401 is straight failure as it has responded and that is what we are
checking for.

403 is failure as the server is refusing to return what it was asked to return.

We give 501 another chance in the same way we treat 500 as some servers just
break on HEAD but do work on GET.