Removed the option for curators to generate a new CSV export.
As the BioCatalogue takes such a long time to generate the CSV export we have removed the option for curators to do so and instead allow curators only to download a pre-generated one.
There is now a download page that tells them when the last export was generated and how to request a new export be generated if it is not recent enough.