Fixed deactivated user issue
When you browse a users profile after they have been deactivated the ‘service responsible’ and ‘favourites’ tab throw a ‘user not found’ related error. I’ve set it so when it’s looking up the user record; it searches for any user, not just ones that haven’t been deactivated.
It may be correct to have all tabs not show if the user is deactivated — the current configuration is that they are all still visible.