Export services in .xls format  and initial work on importing xls files
The CSV export is not satisfactory in use with Rightfield so I've prepared an xls export that does the same function.
The .xls file contains all sheets in one workbook rather than 3 csv files and currently does this 'on-the-fly'.
We will need to batch prepare them.

The import uses the 'roo' gem and is incomplete at this stage